Fix warming

+++ Source/Shared/SafeStringCopy.hxx
@@ -15,7 +15,7 @@ namespace universelan::util {
 	}
 
 	inline bool safe_copy_str_n(const char* const data, char* buffer, uint32_t buffer_len) {
-		uint32_t data_size = std::strlen(data);
+		uint32_t data_size = (uint32_t)std::strlen(data);
 		uint32_t copy_size = std::min((uint32_t)(data_size + 1), buffer_len);
 
 		std::copy_n(data, copy_size, buffer);
